types: дескриптор 64 байта, оси и шаги uint32_t
Ось в uint16_t не вмещала словарь языковой модели: у Qwen2.5 их 151 936.
Оси и шаги стали uint32_t, смещение в арене — uint64_t (потолок арены в
4 ГиБ снят), число элементов по-прежнему uint32_t: парсер теперь
проверяет и его (E0304). Дескриптор вырос с 32 до 64 байт — ровно строка
кэша, без дыр, хвост нулевой. Формат .s3b 0.2: модули 0.1 не читаются.

+++ include/smpc3/ast.h
@@ -58,7 +58,7 @@ typedef struct SmpAstTensor {
     SmpSpan  type_span;
     SmpDType dtype;
     uint32_t rank;
-    uint16_t dims[SMP_MAX_RANK];
+    uint32_t dims[SMP_MAX_RANK];
 
     bool        has_index; /* был ли [...]                                   */
     SmpSpan     index_span;
